# internal/api/docs/swagger.yaml
basePath: /
definitions:
...
  ErrorResponse:
    properties:
      code:
        type: integer
      message:
        type: string
      type:
        type: string
    type: object
  SwaggErrDetailsResp:
    properties:
      error:
        $ref: '#/definitions/ErrorResponse'
    type: object
...
  llm.AiTranscriberResp:
    properties:
      content:
        type: string
    type: object
  llm.AiTranslatorData:
    properties:
      context:
        type: string
      input_text:
        type: string
      source_lang:
        enum:
        - auto
        - vi
        - en
        - th
        - zh
        - ja
        - ko
        - fr
        - de
        - es
        type: string
      target_lang:
        enum:
        - vi
        - en
        - th
        - zh
        - ja
        - ko
        - fr
        - de
        - es
        type: string
    required:
    - input_text
    - source_lang
    - target_lang
    type: object
  llm.AiTranslatorResp:
    properties:
      content:
        type: string
    type: object
  llm.ChatToArticleData:
    properties:
      chat_log:
        type: string
      target_lang:
        enum:
        - en
        - vi
        - fr
        - de
        - es
        - ja
        - ko
        - zh
        type: string
    required:
    - chat_log
    - target_lang
    type: object
  llm.ChatToArticleResp:
    properties:
      content:
        type: string
    type: object
  llm.EmailGeneratorData:
    properties:
      language:
        enum:
        - vi
        - en
        - th
        - zh
        - ja
        - ko
        type: string
      tone:
        enum:
        - professional
        - formal
        - casual
        - friendly
        - urgent
        - apologetic
        type: string
      topic:
        type: string
    required:
    - language
    - tone
    - topic
    type: object
  llm.EmailGeneratorResp:
    properties:
      body:
        type: string
      subject:
        type: string
    type: object
  llm.MarkdownFormatData:
    properties:
      markdown:
        type: string
    required:
    - markdown
    type: object
  llm.MarkdownFormatResp:
    properties:
      content:
        type: string
    type: object
  llm.OcrResp:
    properties:
      content:
        type: string
    type: object
  llm.SmartChatReplyData:
    properties:
      mime_type:
        type: string
      style:
        enum:
        - professional
        - friendly
        - flirty
        - humorous
        - witty
        - mimic
        - empathetic
        - direct
        type: string
      target_lang:
        enum:
        - vi
        - en
        - th
        - zh
        - ja
        - ko
        type: string
      text_input:
        type: string
    required:
    - style
    - target_lang
    type: object
  llm.SmartChatReplyResp:
    properties:
      suggestions:
        items:
          type: string
        type: array
    type: object
  llm.TextSummarizerData:
    properties:
      mode:
        enum:
        - shortest
        - key_points
        - concise
        - detailed
        - professional
        type: string
      target_lang:
        enum:
        - vi
        - en
        - th
        - zh
        - ja
        - ko
        type: string
      text:
        type: string
    required:
    - mode
    - target_lang
    - text
    type: object
  llm.TextSummarizerResp:
    properties:
      content:
        type: string
    type: object
...
info:
  contact:
    email: vuduongcalvin@gmail.com
    name: DuongVu
    url: http://www.swagger.io/support
  description: This is a sample server Core server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: GoCore Example API
  version: "1.0"
paths:
  /api/v1/llm/ai-transcriber:
    post:
      consumes:
      - multipart/form-data
      description: Transcribes an audio/video file using Whisper and translates the
        result to the target language
      operationId: llmAiTranscriber
      parameters:
      - description: Audio/video file to transcribe
        in: formData
        name: file
        required: true
        type: file
      - description: Input language (auto, vi, en, th, zh, ja, ko, fr, de, es)
        in: formData
        name: input_lang
        required: true
        type: string
      - description: Output language (vi, en, th, zh, ja, ko, fr, de, es)
        in: formData
        name: output_lang
        required: true
        type: string
      - description: Output format (text, json, srt)
        in: formData
        name: format
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/llm.AiTranscriberResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/SwaggErrDetailsResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/SwaggErrDetailsResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/SwaggErrDetailsResp'
      security:
      - BearerToken: []
      summary: Transcribe and translate audio/video file
      tags:
      - llm
  /api/v1/llm/ai-translator:
    post:
      consumes:
      - application/json
      description: Translates input text from a source language to a target language
        using a professional AI translator
      operationId: llmAiTranslator
      parameters:
      - description: AiTranslatorData
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/llm.AiTranslatorData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/llm.AiTranslatorResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/SwaggErrDetailsResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/SwaggErrDetailsResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/SwaggErrDetailsResp'
      security:
      - BearerToken: []
      summary: Translate text
      tags:
      - llm
  /api/v1/llm/chat-to-article:
    post:
      consumes:
      - application/json
      description: Takes a raw chat log between a User and AI Assistant and converts
        it into a structured technical article
      operationId: llmChatToArticle
      parameters:
      - description: ChatToArticleData
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/llm.ChatToArticleData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/llm.ChatToArticleResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/SwaggErrDetailsResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/SwaggErrDetailsResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/SwaggErrDetailsResp'
      security:
      - BearerToken: []
      summary: Convert chat log to technical article
      tags:
      - llm
  /api/v1/llm/email-generator:
    post:
      consumes:
      - application/json
      description: Generates a professional email based on topic, tone, and target
        language
      operationId: llmEmailGenerator
      parameters:
      - description: EmailGeneratorData
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/llm.EmailGeneratorData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/llm.EmailGeneratorResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/SwaggErrDetailsResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/SwaggErrDetailsResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/SwaggErrDetailsResp'
      security:
      - BearerToken: []
      summary: Generate professional email
      tags:
      - llm
  /api/v1/llm/markdown-format:
    post:
      consumes:
      - application/json
      description: Reformats raw Markdown text following CommonMark / GitHub Flavored
        Markdown best practices
      operationId: llmMarkdownFormat
      parameters:
      - description: MarkdownFormatData
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/llm.MarkdownFormatData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/llm.MarkdownFormatResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/SwaggErrDetailsResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/SwaggErrDetailsResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/SwaggErrDetailsResp'
      security:
      - BearerToken: []
      summary: Format raw Markdown text
      tags:
      - llm
  /api/v1/llm/ocr:
    post:
      consumes:
      - multipart/form-data
      description: Extracts text from an image using a vision-capable LLM
      operationId: llmOcr
      parameters:
      - description: Image file (jpeg, png, gif, webp)
        in: formData
        name: image
        required: true
        type: file
      - description: Output format (text, markdown, html, json)
        in: formData
        name: format
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/llm.OcrResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/SwaggErrDetailsResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/SwaggErrDetailsResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/SwaggErrDetailsResp'
      security:
      - BearerToken: []
      summary: Extract text from image (OCR)
      tags:
      - llm
  /api/v1/llm/smart-chat-reply:
    post:
      consumes:
      - application/json
      - multipart/form-data
      description: Analyzes a conversation (text or screenshot) and returns 5 reply
        suggestions
      operationId: llmSmartChatReply
      parameters:
      - description: SmartChatReplyData (JSON for text input)
        in: body
        name: request
        schema:
          $ref: '#/definitions/llm.SmartChatReplyData'
      - description: Chat screenshot (multipart)
        in: formData
        name: image
        type: file
      - description: Target language (vi, en, th, zh, ja, ko)
        in: formData
        name: target_lang
        type: string
      - description: Reply style (professional, friendly, flirty, humorous, witty,
          mimic, empathetic, direct)
        in: formData
        name: style
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/llm.SmartChatReplyResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/SwaggErrDetailsResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/SwaggErrDetailsResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/SwaggErrDetailsResp'
      security:
      - BearerToken: []
      summary: Generate smart chat reply suggestions
      tags:
      - llm
  /api/v1/llm/text-summarizer:
    post:
      consumes:
      - application/json
      description: Summarizes and translates text based on the selected mode and target
        language
      operationId: llmTextSummarizer
      parameters:
      - description: TextSummarizerData
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/llm.TextSummarizerData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/llm.TextSummarizerResp'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/SwaggErrDetailsResp'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/SwaggErrDetailsResp'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/SwaggErrDetailsResp'
      security:
      - BearerToken: []
      summary: Summarize and translate text
      tags:
      - llm
...
schemes:
- http
- https
securityDefinitions:
  BearerToken:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"